Refactor to separate database querying code
